+++ apps/nextjs-website/react-components/components/Footer/assets/Assets.helpers.ts
@@ -1,0 +1,6 @@
+import { styled } from "@mui/material";
+
+export const StyledSvg = styled('svg')({
+  display: 'inline-block',
+  userSelect: 'none',
+});
