--- apps/nextjs-website/react-components/components/Footer/assets/LogoPagoPACompany/LogoPagoPACompany.tsx
@@ -1,78 +1,0 @@
-import { styled } from '@mui/material';
-import { theme } from '@pagopa/mui-italia';
-
-const StyledSvg = styled('svg')({
-  display: 'inline-block',
-  userSelect: 'none',
-});
-
-type LogoPACompanyColors = 'default' | 'light' | 'dark';
-type LogoPACompanyVariants = 'default' | 'flat';
-
-interface LogoPagoPACompanyProps {
-  /** Provides a human-readable title for the element that contains it. */
-  title?: string;
-  /** Width of the component. Height is set automatically. */
-  size?: number;
-  /** The color of the component. */
-  color?: LogoPACompanyColors;
-  variant?: LogoPACompanyVariants;
-}
-
-const colorMap = {
-  default: theme.palette.secondary.main,
-  dark: theme.palette.common.black,
-  light: theme.palette.common.white,
-};
-
-export const LogoPagoPACompany = ({
-  title = 'PagoPA',
-  size = 120,
-  color = 'default',
-  variant = 'default',
-}: LogoPagoPACompanyProps): JSX.Element => (
-  <StyledSvg
-    viewBox="0 0 119 33"
-    focusable="false"
-    aria-labelledby="logo-pagoPA-company-titleID"
-    role="img"
-    sx={{
-      width: size,
-      fill: colorMap[color],
-    }}
-  >
-    {variant === 'default' ? <DefaultVariant /> : <FlatVariant />}
-
-    <title id="logo-pagopa-company-title">{title}</title>
-  </StyledSvg>
-);
-
-const DefaultVariant = () => (
-  <>
-    <path
-      opacity=".4"
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M114.338 25.099h3.718V2.324h-3.718v2.97a11.347 11.347 0 0 0-7.669-2.97 11.375 11.375 0 0 0-9.528 5.15 11.334 11.334 0 0 1 1.86 6.237 11.33 11.33 0 0 1-1.86 6.238 11.377 11.377 0 0 0 9.528 5.15c2.954 0 5.645-1.125 7.669-2.97v2.97ZM99 13.71a7.67 7.67 0 1 1 15.338 0 7.67 7.67 0 0 1-15.338 0ZM3.718 5.294v-2.97H0v29.282h3.718v-9.477a11.345 11.345 0 0 0 7.67 2.97c3.985 0 7.493-2.049 9.527-5.15a11.334 11.334 0 0 1-1.859-6.238c0-2.303.684-4.446 1.86-6.237a11.377 11.377 0 0 0-9.529-5.15 11.346 11.346 0 0 0-7.669 2.97Zm0 8.417a7.669 7.669 0 1 1 15.338 0 7.669 7.669 0 0 1-15.338 0Z"
-    />
-    <path
-      opacity=".7"
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M41.83 5.304v-2.98h-3.718v2.97a11.346 11.346 0 0 0-7.669-2.97c-6.288 0-11.387 5.098-11.387 11.387 0 6.29 5.099 11.388 11.387 11.388 2.954 0 5.646-1.125 7.67-2.97v2.97h3.718v-2.925a11.39 11.39 0 0 1-3.719-8.435c0-3.342 1.434-6.348 3.719-8.435Zm-19.055 8.407a7.669 7.669 0 1 1 15.338 0 7.669 7.669 0 0 1-15.338 0ZM79.944 5.293V2.324h-3.719v2.97a11.358 11.358 0 0 1 3.719 8.417c0 3.335-1.434 6.335-3.719 8.418v9.477h3.719v-9.477a11.346 11.346 0 0 0 7.669 2.97C93.902 25.099 99 20 99 13.71S93.902 2.324 87.613 2.324a11.346 11.346 0 0 0-7.669 2.97Zm0 8.418a7.669 7.669 0 1 1 15.338 0 7.669 7.669 0 0 1-15.338 0Z"
-    />
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M60.887 0v5.293l.157-.14a11.344 11.344 0 0 1 7.513-2.83c6.288 0 11.387 5.1 11.387 11.388 0 6.29-5.099 11.388-11.388 11.388a11.345 11.345 0 0 1-7.668-2.97v3.02c0 2.9-1.16 5.627-3.196 7.7l-.15.151-2.604-2.66c1.383-1.36 2.179-3.123 2.229-5.002l.002-.189v-2.975A11.333 11.333 0 0 1 49.5 25.15c-6.289 0-11.387-5.11-11.387-11.41 0-6.303 5.098-11.411 11.387-11.411 2.954 0 5.645 1.127 7.67 2.975V0h3.717ZM49.5 6.055c-4.235 0-7.669 3.44-7.669 7.684s3.434 7.685 7.67 7.685c4.235 0 7.668-3.44 7.668-7.685 0-4.244-3.433-7.684-7.669-7.684Zm11.388 7.656a7.669 7.669 0 1 1 15.338 0 7.669 7.669 0 0 1-15.338 0Z"
-    />
-  </>
-);
-
-const FlatVariant = () => (
-  <>
-    <path d="M119 2.343v22.956h-3.748v-2.993a11.437 11.437 0 0 1-7.73 2.993c-3.66 0-6.92-1.713-9.022-4.38a12.321 12.321 0 0 0 1.875-4.147 7.73 7.73 0 1 0 0-5.903 12.337 12.337 0 0 0-1.874-4.145 11.455 11.455 0 0 1 9.021-4.381c2.977 0 5.69 1.133 7.73 2.993V2.343H119ZM3.748 2.343v2.993a11.436 11.436 0 0 1 7.73-2.993c3.66 0 6.92 1.712 9.022 4.38a12.35 12.35 0 0 0-1.875 4.147 7.732 7.732 0 1 0 0 5.902 12.347 12.347 0 0 0 1.874 4.147 11.457 11.457 0 0 1-9.02 4.38c-2.978 0-5.691-1.134-7.731-2.993v9.552H0V2.343h3.748Z" />
-    <path d="M42.165 2.343v1.763a12.42 12.42 0 0 0-4.332 6.764 7.73 7.73 0 1 0 0 5.902 12.415 12.415 0 0 0 4.332 6.764V25.3h-3.748v-2.993a11.436 11.436 0 0 1-7.73 2.993c-6.34 0-11.478-5.139-11.478-11.478 0-6.34 5.139-11.478 11.478-11.478 2.977 0 5.69 1.133 7.73 2.993V2.343h3.748ZM80.583 2.343v2.992a11.437 11.437 0 0 1 7.73-2.992c6.34 0 11.478 5.139 11.478 11.478 0 6.34-5.139 11.478-11.478 11.478-2.978 0-5.69-1.133-7.73-2.993v9.552h-3.748v-8.322a12.42 12.42 0 0 0 4.332-6.764 7.73 7.73 0 1 0 0-5.903 12.415 12.415 0 0 0-4.332-6.764V2.343h3.748Z" />
-    <path d="M61.374 0v5.336a11.437 11.437 0 0 1 7.73-2.993c6.34 0 11.479 5.139 11.479 11.478 0 6.34-5.14 11.478-11.479 11.478-2.977 0-5.69-1.134-7.73-2.993v3.044c0 2.924-1.17 5.672-3.221 7.762l-.152.152-2.624-2.682c1.394-1.37 2.196-3.147 2.246-5.042l.003-.19v-2.999c-2.04 1.863-4.753 3-7.73 3-6.34 0-11.479-5.15-11.479-11.502s5.14-11.502 11.479-11.502c2.977 0 5.69 1.136 7.73 3V0h3.748ZM49.896 6.103c-4.27 0-7.73 3.468-7.73 7.746 0 4.278 3.46 7.746 7.73 7.746 4.269 0 7.73-3.468 7.73-7.746 0-4.278-3.461-7.746-7.73-7.746Zm19.208-.012a7.73 7.73 0 1 0 0 15.46 7.73 7.73 0 0 0 0-15.46Z" />
-  </>
-);
