--- apps/nextjs-website/react-components/components/Footer/assets/LogoPagoPACompany/index.ts
@@ -1,1 +1,0 @@
-export * from './LogoPagoPACompany';
