+++ apps/nextjs-website/react-components/components/Footer/assets/LogoPagoPACompany/partials/FlatVariant.tsx
@@ -1,0 +1,7 @@
+export const FlatVariant = () => (
+  <>
+    <path d="M119 2.343v22.956h-3.748v-2.993a11.437 11.437 0 0 1-7.73 2.993c-3.66 0-6.92-1.713-9.022-4.38a12.321 12.321 0 0 0 1.875-4.147 7.73 7.73 0 1 0 0-5.903 12.337 12.337 0 0 0-1.874-4.145 11.455 11.455 0 0 1 9.021-4.381c2.977 0 5.69 1.133 7.73 2.993V2.343H119ZM3.748 2.343v2.993a11.436 11.436 0 0 1 7.73-2.993c3.66 0 6.92 1.712 9.022 4.38a12.35 12.35 0 0 0-1.875 4.147 7.732 7.732 0 1 0 0 5.902 12.347 12.347 0 0 0 1.874 4.147 11.457 11.457 0 0 1-9.02 4.38c-2.978 0-5.691-1.134-7.731-2.993v9.552H0V2.343h3.748Z" />
+    <path d="M42.165 2.343v1.763a12.42 12.42 0 0 0-4.332 6.764 7.73 7.73 0 1 0 0 5.902 12.415 12.415 0 0 0 4.332 6.764V25.3h-3.748v-2.993a11.436 11.436 0 0 1-7.73 2.993c-6.34 0-11.478-5.139-11.478-11.478 0-6.34 5.139-11.478 11.478-11.478 2.977 0 5.69 1.133 7.73 2.993V2.343h3.748ZM80.583 2.343v2.992a11.437 11.437 0 0 1 7.73-2.992c6.34 0 11.478 5.139 11.478 11.478 0 6.34-5.139 11.478-11.478 11.478-2.978 0-5.69-1.133-7.73-2.993v9.552h-3.748v-8.322a12.42 12.42 0 0 0 4.332-6.764 7.73 7.73 0 1 0 0-5.903 12.415 12.415 0 0 0-4.332-6.764V2.343h3.748Z" />
+    <path d="M61.374 0v5.336a11.437 11.437 0 0 1 7.73-2.993c6.34 0 11.479 5.139 11.479 11.478 0 6.34-5.14 11.478-11.479 11.478-2.977 0-5.69-1.134-7.73-2.993v3.044c0 2.924-1.17 5.672-3.221 7.762l-.152.152-2.624-2.682c1.394-1.37 2.196-3.147 2.246-5.042l.003-.19v-2.999c-2.04 1.863-4.753 3-7.73 3-6.34 0-11.479-5.15-11.479-11.502s5.14-11.502 11.479-11.502c2.977 0 5.69 1.136 7.73 3V0h3.748ZM49.896 6.103c-4.27 0-7.73 3.468-7.73 7.746 0 4.278 3.46 7.746 7.73 7.746 4.269 0 7.73-3.468 7.73-7.746 0-4.278-3.461-7.746-7.73-7.746Zm19.208-.012a7.73 7.73 0 1 0 0 15.46 7.73 7.73 0 0 0 0-15.46Z" />
+  </>
+);
