+++ apps/nextjs-website/react-components/components/Footer/assets/LogoPagoPACompany/partials/index.ts
@@ -1,0 +1,2 @@
+export * from './DefaultVariant';
+export * from './FlatVariant';
